+++ src/collective/documentgenerator/setuphandlers.py
@@ -10,6 +10,8 @@ import logging
 
 from Products.CMFPlone.utils import safe_unicode
 
+from config import set_oo_port
+
 logger = logging.getLogger('collective.documentgenerator')
 
 
@@ -21,6 +23,7 @@ def post_install(context):
     """Post install script"""
     if isNotCurrentProfile(context):
         return
+    set_oo_port()
 
 
 def install_demo(context):